Warning shutup
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@1474 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/binaries/web/cgi-src/math_sum/math_sum.cpp
@@ -8,7 +8,7 @@ int isNumber(char* s)
 {
   if(!s)
     return 0;
-  for(int i=0; i<strlen(s);i++)
+  for(u_long i=0; i<strlen(s);i++)
     if(!isdigit(s[i]))
       return 0;
   return 1;
